Decouple RPC/Graph views for collection
